# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: DST Admin API
  version: 2.0.0-alpha.1
  description: >-
    Contract for the authenticated DST Admin management API. JSON fields use
    camelCase. Authenticated state-changing requests require X-CSRF-Token and
    Idempotency-Key headers. Clients send X-DST-Runtime-Target with local or an
    agent-prefixed target ID; unsupported remote domain operations fail closed.
servers:
  - url: /api/v2
...
paths:
...
  /rooms/{roomId}/players:
    parameters: [{ $ref: "#/components/parameters/RoomID" }]
    get:
      tags: [Players]
      operationId: listRoomPlayers
      summary: List persisted player observations and current access state
      parameters:
        - { name: query, in: query, schema: { type: string, maxLength: 128 } }
        - { name: status, in: query, schema: { type: string, enum: [online, offline] } }
        - { name: worldId, in: query, schema: { type: string } }
        - { name: prefab, in: query, schema: { type: string, maxLength: 128 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
      responses:
        "200": { description: Player history and status summary, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerListEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    get:
      tags: [Players]
      operationId: getRoomPlayer
      summary: Read one persisted player observation
      responses:
        "200": { description: Player details, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/actions/refresh:
    parameters: [{ $ref: "#/components/parameters/RoomID" }]
    post:
      tags: [Players, Jobs]
      operationId: refreshRoomPlayers
      summary: Probe each world and persist its current player snapshot
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerRefreshInput" } } } }
      responses:
        "202": { description: Player refresh job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{roomId}/players/{playerId}/actions/kick:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: kickRoomPlayer
      summary: Kick an online player from the world where they were observed
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player kick job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/ban:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: banRoomPlayer
      summary: Add a player to the room blocklist and notify a running shard
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player ban job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/unban:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: unbanRoomPlayer
      summary: Remove a player from the room blocklist and notify a running shard
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player unban job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/announce:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: announceToRoomPlayer
      summary: Broadcast a marked reminder on the online player's current shard
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player announcement job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/kill:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: killRoomPlayer
      summary: Kill an online player in the world where they were observed
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player kill job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/god-mode:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: setRoomPlayerGodMode
      summary: Enable or disable invincibility for an online player
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player god mode job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/creative-mode:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: setRoomPlayerCreativeMode
      summary: Enable or disable free-build mode for an online player
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player creative mode job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/resurrect:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: resurrectRoomPlayer
      summary: Resurrect an online ghost player
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player resurrection job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/players/{playerId}/actions/change-character:
    parameters: [{ $ref: "#/components/parameters/RoomID" }, { $ref: "#/components/parameters/PlayerID" }]
    post:
      tags: [Players, Jobs]
      operationId: resetRoomPlayerCharacter
      summary: Despawn an online player so they can choose another character
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/PlayerActionInput" } } } }
      responses:
        "202": { description: Player character reset job accepted, content: { application/json: { schema: { $ref: "#/components/schemas/JobEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dst_admin_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A unique operation key. Reusing it with the same request replays the original response; reusing it with different JSON content returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$'
    RoomID:
      name: roomId
      in: path
      required: true
      schema:
        type: string
...
    PlayerID:
      name: playerId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{1,128}$'
...
  responses:
...
    Conflict:
      description: The request conflicts with current setup state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: The request is valid JSON but violates a domain constraint
      content:
        application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
    APIError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
    ErrorEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: "null"
        error:
          $ref: "#/components/schemas/APIError"
        meta:
          $ref: "#/components/schemas/Meta"
...
    Player:
      type: object
      additionalProperties: false
      required: [id, roomId, worldId, worldName, name, online, banned, admin, age, firstSeenAt, lastSeenAt, statusChangedAt, lastRefreshedAt, fields]
      properties:
        id: { type: string, pattern: '^[A-Za-z0-9_-]{1,128}$' }
        roomId: { type: string }
        worldId: { type: string }
        worldName: { type: string }
        name: { type: string }
        prefab: { type: string }
        online: { type: boolean }
        banned: { type: boolean }
        banReason: { type: string, maxLength: 300 }
        bannedAt: { type: string, format: date-time }
        banExpiresAt: { type: string, format: date-time }
        admin: { type: boolean }
        age: { type: integer, minimum: 0 }
        netId: { type: string }
        netScore: { type: integer, minimum: 0 }
        performance: { type: integer, deprecated: true, description: Legacy field retained for read compatibility; use netScore }
        healthPercent: { type: number }
        hungerPercent: { type: number }
        sanityPercent: { type: number }
        temperature: { type: number }
        moisture: { type: number }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        statusChangedAt: { type: string, format: date-time }
        lastRefreshedAt: { type: string, format: date-time }
        fields:
          type: object
          additionalProperties: { $ref: "#/components/schemas/PlayerFieldState" }
    PlayerDataSource:
      type: string
      enum: [customcommands, console-fallback, native-log]
    PlayerFreshnessStatus:
      type: string
      enum: [live, stale, unavailable]
    PlayerFieldState:
      type: object
      additionalProperties: false
      required: [source, status]
      properties:
        source: { $ref: "#/components/schemas/PlayerDataSource" }
        observedAt: { type: string, format: date-time }
        status: { $ref: "#/components/schemas/PlayerFreshnessStatus" }
    PlayerEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/Player" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    PlayerList:
      type: object
      additionalProperties: false
      required: [items, total, online, offline, banned, limit, offset]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Player" } }
        total: { type: integer, minimum: 0 }
        online: { type: integer, minimum: 0 }
        offline: { type: integer, minimum: 0 }
        banned: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        offset: { type: integer, minimum: 0 }
        lastRefreshedAt: { type: string, format: date-time }
    PlayerListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/PlayerList" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    PlayerActionInput:
      type: object
      additionalProperties: false
      properties:
        worldId: { type: string }
        message: { type: string, maxLength: 300 }
        confirmation: { type: string }
        reason: { type: string, maxLength: 300 }
        duration: { type: string, enum: [1h, 6h, 12h, 1d, 3d, 7d, 30d, permanent] }
        enabled: { type: boolean }
    PlayerRefreshInput:
      type: object
      additionalProperties: false
      properties:
        worldIds: { type: array, maxItems: 64, uniqueItems: true, items: { type: string } }
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
    JobOutcome:
      type: string
      enum: [pending, full, partial, none]
    JobError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    JobTarget:
      type: object
      additionalProperties: false
      required: [id, targetId, name, status]
      properties:
        id:
          type: integer
          format: int64
        targetId:
          type: string
        name:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        message:
          type: string
        error:
          $ref: "#/components/schemas/JobError"
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
    Job:
      type: object
      additionalProperties: false
      required:
        - id
        - kind
        - status
        - outcome
        - progress
        - cancelRequested
        - createdAt
        - targets
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        outcome:
          $ref: "#/components/schemas/JobOutcome"
        roomId:
          type: string
        worldId:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        error:
          $ref: "#/components/schemas/JobError"
        cancelRequested:
          type: boolean
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        targets:
          type: array
          items:
            $ref: "#/components/schemas/JobTarget"
...
    JobEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/Job"
        error:
          type: "null"
        meta:
          $ref: "#/components/schemas/Meta"
...
security:
  - sessionCookie: []